setup.sh file edited for giving support to other distro packagemanger for mysql installation

+++ setup.sh
@@ -11,7 +11,7 @@ if [ -f "requirements.txt" ]; then
 	echo "Installing python packages from requirements.txt file......"
 	if command_exists pip; then
 		pip install -r requirements.txt
-	elfi command_exists pip3; then
+	elif command_exists pip3; then
 		pip3 install -r requirements.txt
 	else
 		echo "Pip not found.......Install pip or start your virtual environment if needed..........."
